fix(watch): finish portability and arbitration hardening
