+++ internal/watcher/takeover_unix_test.go
@@ -200,9 +200,8 @@ func TestUnixStaleSocketFailsSafely(t *testing.T) {
 	}
 }
 
-// macOS bounds sockaddr_un sun_path to 104 bytes; os.TempDir() alone is often
-// most of that, so the derived filename has to be compact. Rising over the
-// limit here is a bind: invalid argument failure on mac CI, so pin it.
+// macOS bounds sockaddr_un sun_path to 104 bytes; the fixed short directory and
+// compact derived filename keep binding valid regardless of TMPDIR.
 func TestUnixSocketPathStaysWithinSockaddrUnLimit(t *testing.T) {
 	home := t.TempDir()
 	if path := takeoverSocketPath(home, "deadbeef"+strings.Repeat("a", 24)); len(path) >= sockaddrUnLimit {
@@ -210,15 +209,16 @@ func TestUnixSocketPathStaysWithinSockaddrUnLimit(t *testing.T) {
 	}
 }
 
-func TestUnixSocketPathFallsBackFromAnOverlongTempDir(t *testing.T) {
+func TestUnixSocketPathDoesNotDependOnTempDir(t *testing.T) {
 	home := t.TempDir()
+	want := takeoverSocketPath(home, genB)
 	t.Setenv("TMPDIR", filepath.Join(string(filepath.Separator), strings.Repeat("t", 120)))
 	path := takeoverSocketPath(home, genB)
 	if len(path) >= sockaddrUnLimit {
 		t.Fatalf("takeover socket path %q is %d bytes, want < %d", path, len(path), sockaddrUnLimit)
 	}
-	if !strings.HasPrefix(path, string(filepath.Separator)+"tmp"+string(filepath.Separator)) {
-		t.Fatalf("takeover socket path %q did not use the short fallback directory", path)
+	if path != want {
+		t.Fatalf("takeover socket path changed with TMPDIR: before %q, after %q", want, path)
 	}
 }
